backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
openai==1.3.7
twilio==8.10.0
selenium==4.15.2
beautifulsoup4==4.12.2
pandas==2.1.3
PyPDF2==3.0.1
python-jose==3.3.0
bcrypt==4.1.1
email-validator==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4